+++ lib/esereader.h
@@ -30,21 +30,21 @@ class ESEReader {
   /// @brief Reset the reader
   virtual void reset ();
 
-  virtual bool      prepare ()                = 0;
-  virtual ESEBuffer getBuffer (uint32_t size) = 0;
+  virtual bool      prepare ()              = 0;
+  virtual ESEBuffer getBuffer (size_t size) = 0;
 
-  int32_t        readSize () { return m_readSize; }
+  size_t         readSize () { return m_readSize; }
   virtual size_t streamSize () = 0;
   int32_t        streamPosition () { return m_streamPosition; }
-  uint32_t       bufferReadLength () { return m_bufferReadLength; }
-  void           setBufferReadLength (uint32_t bufferReadLength) { m_bufferReadLength = bufferReadLength; }
+  size_t         bufferReadLength () { return m_bufferReadLength; }
+  void           setBufferReadLength (size_t bufferReadLength) { m_bufferReadLength = bufferReadLength; }
 
   virtual bool isEOS () = 0;
 
   protected:
   int32_t   m_streamPosition;
-  int32_t   m_bufferSize;
+  size_t    m_bufferSize;
   size_t    m_readSize;
-  uint32_t  m_bufferReadLength;
+  size_t    m_bufferReadLength;
   ESEBuffer m_buffer;
 };
